Port to rendering data out of component
